rename: 폰트 설정파일 경로수정 #4
