+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
-import notoSansKr from '@/styles/fonts';
+import notoSansKr from '@/styles/font';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
